Vunerabilites and restructure started
