+++ package.json
@@ -15,38 +15,38 @@
   },
   "dependencies": {
     "@diracleo/vue-scrubbable-video": "0.0.8",
-    "@fortawesome/fontawesome-svg-core": "^1.2.30",
-    "@fortawesome/free-brands-svg-icons": "^5.14.0",
-    "@fortawesome/free-solid-svg-icons": "^5.14.0",
-    "@fortawesome/vue-fontawesome": "^0.1.10",
-    "core-js": "^3.6.5",
-    "d3": "^5.16.0",
-    "dns": "^0.2.2",
-    "gsap": "^3.6.0",
-    "lint": "^0.7.0",
-    "mobile-device-detect": "^0.4.3",
-    "roughjs": "^4.3.1",
-    "uswds": "^2.8.1",
-    "vue": "^2.6.12",
-    "vue-analytics": "^5.22.1",
-    "vue-browser-detect-plugin": "^0.1.13",
-    "vue-mapbox": "^0.4.1",
-    "vue-mobile-detection": "^1.0.0",
-    "vue-parallaxy": "^1.1.1",
-    "vue-router": "^3.4.3",
-    "vuetify": "^2.4.3",
-    "vuex": "^3.5.1"
+    "@fortawesome/fontawesome-svg-core": "1.2.30",
+    "@fortawesome/free-brands-svg-icons": "5.14.0",
+    "@fortawesome/free-solid-svg-icons": "5.14.0",
+    "@fortawesome/vue-fontawesome": "0.1.10",
+    "core-js": "3.6.5",
+    "d3": "5.16.0",
+    "gsap": "3.6.0",
+    "lint": "0.7.0",
+    "mobile-device-detect": "0.4.3",
+    "roughjs": "4.3.1",
+    "uswds": "2.8.1",
+    "vue": "2.6.12",
+    "vue-analytics": "5.22.1",
+    "vue-browser-detect-plugin": "0.1.13",
+    "vue-mapbox": "0.4.1",
+    "vue-mobile-detection": "1.0.0",
+    "vue-parallaxy": "1.1.1",
+    "vue-router": "3.4.3",
+    "vuetify": "2.4.3",
+    "vuex": "3.5.1",
+    "ws": "^6.2.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.4",
-    "@vue/cli-plugin-eslint": "^4.5.4",
-    "@vue/cli-service": "^4.5.4",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^7.7.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^9.0.3",
-    "vue-svg-loader": "^0.16.0",
-    "vue-template-compiler": "^2.6.12"
+    "@vue/cli-plugin-babel": "4.5.4",
+    "@vue/cli-plugin-eslint": "4.5.4",
+    "@vue/cli-service": "4.5.4",
+    "babel-eslint": "10.1.0",
+    "eslint": "7.7.0",
+    "eslint-plugin-vue": "6.2.2",
+    "node-sass": "4.14.1",
+    "sass-loader": "9.0.3",
+    "vue-svg-loader": "0.16.0",
+    "vue-template-compiler": "2.6.12"
   }
 }
