+++ src/views/Visualization.vue
@@ -1,16 +1,5 @@
 <template>
-  <div id="visualization">
-    <!-- <Intro />
-     -->
-    <Header />
-    <Timing />
-            <Broadscale />
-    <SWE />
-    <SWEanimation />
-<!--     <Broadscale /> -->
-     <!-- <In2021 /> -->
-    <References />
-  </div>
+  <div id="visualization" />
 </template>
 
 <script>
@@ -18,15 +7,14 @@
     export default {
         name: 'Visualization',
         components: {
-          Timing: () => import(/* webpackPreload: true */ /*webpackChunkName: "intro"*/ "./timing/Timing"),
-          Header: () => import(/* webpackPreload: true */ /*webpackChunkName: "Header"*/ "./header/Header"),
-          SWE: () => import(/* webpackPreload: true */ /*webpackChunkName: "swe"*/ "./swe/SWE"),
-          SWEanimation: () => import(/* webpackPreload: true */ /*webpackChunkName: "swea"*/ "./swe/SWEanimation"),
-          //In2021: () => import(/* webpackPreload: true */ /*webpackChunkName: "In2021"*/ "./in2021/In2021"),
-          // Intro: () => import(/* webpackPreload: true */ /*webpackChunkName: "intro"*/ "./intro/Intro"),
-          Broadscale: () => import(/* webpackPreload: true */ /*webpackChunkName: "broadscale"*/ "./1-Broadscale/Broadscale"),
-          References: () => import(/* webpackPreload: true */ /*webpackChunkName: "references"*/ "./4-References/References"),
-          
+          // Timing: () => import(/* webpackPreload: true */ /*webpackChunkName: "intro"*/ "@/components/Timing"),
+          // Splash: () => import(/* webpackPreload: true */ /*webpackChunkName: "Header"*/ "@/components/Splash"),
+          // // SWE: () => import(/* webpackPreload: true */ /*webpackChunkName: "swe"*/ "./swe/SWE"),
+          // SWEanimation: () => import(/* webpackPreload: true */ /*webpackChunkName: "swea"*/ "@/components/SWEanimation"),
+          // //In2021: () => import(/* webpackPreload: true */ /*webpackChunkName: "In2021"*/ "./in2021/In2021"),
+          // // Intro: () => import(/* webpackPreload: true */ /*webpackChunkName: "intro"*/ "./intro/Intro"),
+          // Broadscale: () => import(/* webpackPreload: true */ /*webpackChunkName: "broadscale"*/ "@/components/Broadscale"),
+          // References: () => import(/* webpackPreload: true */ /*webpackChunkName: "references"*/ "@/components/References"), 
         }
     }
 </script>
