--- src/views/swe/SWEanimation.vue
@@ -1,213 +1,0 @@
-<template>
-  <section id="swe-animation">
-    <div class="text-content">
-        <div>
-          <h2 id="no-flow">Changes in snowmelt have downstream consequences</h2>
-          <p> Pepper jack melted cheese feta. Cheesy grin taleggio fromage edam boursin manchego cheese triangles parmesan. Fromage cheese and biscuits say cheese bocconcini gouda lancashire cheese slices ricotta. Rubber cheese melted cheese cheesy grin everyone loves mascarpone.</p><br><br>
-        <svg id="swe-gifs"
-          xmlns="http://www.w3.org/2000/svg"
-          xmlns:xlink="http://www.w3.org/1999/xlink"
-          viewBox="0 0 1000 900"
-          aria-labelledby="page-title page-desc"
-          width="100%"
-          >
-          <title id="page-title">not all snow is flow</title>
-        <desc id="page-desc">An animated one-year time series of snow-water equivalent in  the upper colorado river basin</desc>
-        
-          <image
-            xlink:href="@/assets/videos/swe_2011.gif"
-            height="400px"
-            x="0"
-            y="50"
-          />
-          <image
-            xlink:href="@/assets/videos/swe_2012.gif"
-            height="400px"
-            x="0"
-            y="475"
-          />
-          <g id="gages_2011" >
-            <rect x="0" y="50" width="430" height="400" style="fill: none;stroke: black;stroke-linecap: round;stroke-linejoin: round;stroke-width: 3px"/>
-            <circle fill="#ffe0e0" id="gage_2011_06714800" cx="392.26" cy="134.07" r="2.25" class="gage"/>
-            <circle fill="#ebb5a8" id="gage_2011_0671500" cx="396.26" cy="125.37" r="2.25" class="gage"/>
-            <circle fill="#d9b29a" id="gage_2011_06716100" cx="398.19" cy="124.44" r="2.25" class="gage"/>
-            <circle fill="#d6b399" id="gage_2011_06716500" cx="400" cy="123.53" r="2.25" class="gage"/>
-            <circle fill="#cdb194" id="gage_2011_06717400" cx="406.38" cy="130.22" r="2.25" class="gage"/>
-            <circle fill="#c0ad8d" id="gage_2011_09016500" cx="373.25" cy="76.67" r="2.25" class="gage"/>
-            <circle fill="#99937a" id="gage_2011_09019000" cx="376.12" cy="72.29" r="2.25" class="gage"/>
-            <circle fill="#8d8975" id="gage_2011_09019500" cx="372.64" cy="75.4" r="2.25" class="gage"/>
-            <circle fill="#757167" id="gage_2011_09022000" cx="402" cy="112.64" r="2.25" class="gage"/>
-            <circle fill="#757167" id="gage_2011_09024000" cx="385.21" cy="105.3" r="2.25" class="gage"/>
-            <circle fill="#66655f" id="gage_2011_09025000" cx="384.33" cy="102.56" r="2.25" class="gage"/>
-            <circle fill="#62615d" id="gage_2011_09025300" cx="379.39" cy="106.68" r="2.25" class="gage"/>
-            <circle fill="#62615d" id="gage_2011_09025400" cx="380.12" cy="102.67" r="2.25" class="gage"/>
-            <circle fill="#62615d" id="gage_2011_09026500" cx="374.66" cy="103.86" r="2.25" class="gage"/>
-            <circle fill="#555552" id="gage_2011_09032000" cx="386.42" cy="98.57" r="2.25" class="gage"/>
-            <circle fill="#555552" id="gage_2011_09032100" cx="388.59" cy="93.76" r="2.25" class="gage"/>
-            <circle fill="#555552" id="gage_2011_09033300" cx="377.9" cy="90.81" r="2.25" class="gage"/>
-            <circle fill="#555552" id="gage_2011_09034900" cx="371.54" cy="124.02" r="2.25" class="gage"/>
-            <circle fill="#ebb5a8" id="gage_2011_035500" cx="369.29" cy="121.49" r="2.25" class="gage"/>
-            <circle fill="#ebb5a8" id="gage_2011_394308105413800" cx="393.56" cy="129.8" r="2.25" class="gage"/>
-          </g>
-            <g id="gages_2012" >
-            <rect x="0" y="475" width="430" height="400" style="fill: none;stroke: black;stroke-linecap: round;stroke-linejoin: round;stroke-width: 3px"/>
-            <circle fill="#ffe0e0" id="gage_2012_06714800" cx="392.26" cy="559.07" r="2.25" class="gage"/>
-            <circle fill="#ebb5a8" id="gage_2012_0671500" cx="396.26" cy="559.37" r="2.25" class="gage"/>
-            <circle fill="#d9b29a" id="gage_2012_06716100" cx="398.19" cy="5584.44" r="2.25" class="gage"/>
-            <circle fill="#d6b399" id="gage_2012_06716500" cx="400" cy="557.53" r="2.25" class="gage"/>
-            <circle fill="#cdb194" id="gage_2012_06717400" cx="406.38" cy="555.22" r="2.25" class="gage"/>
-            <circle fill="#c0ad8d" id="gage_2012_09016500" cx="373.25" cy="501.67" r="2.25" class="gage"/>
-            <circle fill="#99937a" id="gage_2012_09019000" cx="376.12" cy="497.29" r="2.25" class="gage"/>
-            <circle fill="#8d8975" id="gage_2012_09019500" cx="372.64" cy="500.4" r="2.25" class="gage"/>
-            <circle fill="#757167" id="gage_2012_09022000" cx="402" cy="537.64" r="2.25" class="gage"/>
-            <circle fill="#757167" id="gage_2012_09024000" cx="385.21" cy="530.3" r="2.25" class="gage"/>
-            <circle fill="#66655f" id="gage_2012_09025000" cx="384.33" cy="527.56" r="2.25" class="gage"/>
-            <circle fill="#62615d" id="gage_2012_09025300" cx="379.39" cy="531.68" r="2.25" class="gage"/>
-            <circle fill="#62615d" id="gage_2012_09025400" cx="380.12" cy="527.67" r="2.25" class="gage"/>
-            <circle fill="#62615d" id="gage_2012_09026500" cx="374.66" cy="528.86" r="2.25" class="gage"/>
-            <circle fill="#555552" id="gage_2012_09032000" cx="386.42" cy="523.57" r="2.25" class="gage"/>
-            <circle fill="#555552" id="gage_2012_09032100" cx="388.59" cy="518.76" r="2.25" class="gage"/>
-            <circle fill="#555552" id="gage_2012_09033300" cx="377.9" cy="515.81" r="2.25" class="gage"/>
-            <circle fill="#555552" id="gage_2012_09034900" cx="371.54" cy="549.02" r="2.25" class="gage"/>
-            <circle fill="#ebb5a8" id="gage_2012_035500" cx="369.29" cy="546.49" r="2.25" class="gage"/>
-            <circle fill="#ebb5a8" id="gage_2012_394308105413800" cx="393.56" cy="554.8" r="2.25" class="gage"/>
-          </g>
-          <image
-            xlink:href="@/assets/videos/mmd_2011.gif"
-            height="450px"
-            x="450"
-            y="50"
-          />
-          <image
-            xlink:href="@/assets/videos/mmd_2012.gif"
-            height="450px"
-            x="450"
-            y="475"
-          />
-          <text class="axis-label big-statement" x="500" y="100">high snow (2011)</text>
-        <text class="axis-label big-statement" x="500" y="550">low snow (2012)</text>
-        </svg>
-      </div>
-<div class="text-content"><p>Who moved my cheese boursin babybel. Monterey jack jarlsberg port-salut cheesy feet croque monsieur babybel ricotta mascarpone. Halloumi roquefort stinking bishop say cheese hard cheese rubber cheese mascarpone cheesecake. Babybel the big cheese cow st. agur blue cheese.</p></div>
-    </div>
-  </section>
-</template>
-
-<script>
-import * as d3Base from "d3";
-    export default {
-        name: 'SWE',
-        data() {
-            return {
-              title: process.env.VUE_APP_TITLE,
-              mmd_2011: null,
-              mmd_2012: null,
-            }
-        },
-        computed: {
-            windowHeight: function () {
-                const usgsBannerHeight = 85;
-                return Number(this.$store.state.windowHeight - this.$store.state.warningHeight - usgsBannerHeight) + 'px';
-            }
-        },
-        mounted() {
-          const self = this;
-          this.d3 = Object.assign(d3Base);
-
-        },
-        methods: {
-          loadData() {
-            const self = this;
-            // read in data 
-            let promises = [self.d3.csv(self.publicPath + "gage_mmd_2011.csv", this.d3.autoType),
-            self.d3.csv(self.publicPath + 'gage_mmd_2012.csv'), this.d3.autotype];
-
-            Promise.all(promises).then(self.callback); 
-          },
-          callback(data) {
-            const  self  = this;
-
-            this.mmd_2011 = data[0];
-            this.mmd_2012 = data[1];
-
-            //trigger hydrograph animation
-            self.drawHydrographs();
-
-          },
-          drawHydrographs() {
-            const self = this;
-
-            this.svg  = this.d3.select("#swe-gifs")
-
-            
-          }
-        }
-    }
-</script>
-
-<style scoped lang="scss">
-text.axis-label {
-  fill:black;
-  z-index:1;
-}
-
-#swe-animation-viz {
-    height: 1000px;
-}
-.sketchy {
-    stroke-width: 2px;
-    fill: none;
-    stroke: grey
-
-}
-.sketchy-blue {
-    stroke-width:10px;
-    fill: none;
-    stroke: #0e64bb;
-
-}
-.axes {
-    fill: none;
-    stroke-width: 1px;
-    stroke-dasharray: 2px;
-    stroke: grey;
-}
-#swe-viz {
-    position: relative;
-    width: 80%;
-}
-#no-flow {
-  margin: 50px;
-}
-line, polyline, polygon, path, rect, circle {
-      fill: none;
-      stroke: grey;
-      stroke-linecap: round;
-      stroke-linejoin: round;
-      stroke-miterlimit: 10.00;
-    }
-.gage {
-      //fill: #FF007F;
-      stroke: #FF007F;
-}
-.grey-box {
-      height: 500px;
-}
-h2.big-statement {
-  margin: 20px;
-      color: #0e64bb;
-      fill: white;
-    }
-.left {
- 
-
-}
-.right {
-  width:100%;
-}
-#swe-animation-viz {
-  display:grid;
-  grid-template-columns: 1fr 1fr;
-  grid-template-rows: 1fr;
-}
-</style>
